requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
openai
python-dotenv
psycopg2-binary
requests
